+++ management.php
@@ -28,8 +28,11 @@ include "./utils.php";
             <a class="button" role="button" href="service.php">Service</a><br><br><br>
             <?php
             if ($config["advanced"] == true) {
-                echo '<a class="button" role="button" href="update.php">Update</a><br><br><br>';
                 echo '<a class="button" role="button" href="files.php">Files</a><br><br><br>';
+                echo '<a class="button" role="button" href="update.php">Update</a><br><br><br>';
+                echo '<a class="button" role="button" href="permissions.php">Permissions</a><br><br><br>';
+            } else {
+                echo "<p><i>To access advanced tools, enable 'Advanced Mode' in the controller settings.</i></p>";
             }
             ?>
         </main>